more fixes for MacOS

+++ gtsam_unstable/discrete/examples/schedulingExample.cpp
@@ -224,7 +224,7 @@ void sampleSolutions() {
 
   // now, sample schedules
   for (size_t n = 0; n < 500; n++) {
-    vector<size_t> stats(19, 0);
+    vector<uint64_t> stats(19, 0);
     vector<Values> samples;
     for (size_t i = 0; i < 7; i++) {
       samples.push_back(samplers[i]->sample());
